chore: Big signature and class mismatch pass
signature: 15,267 -> 15,072
class: 1,248 -> 911

+++ Engine/Source/Gx/CGxStateBom.h
@@ -38,7 +38,7 @@ class CGxStateBom {
 struct CGxPushedRenderState {
   EGxRenderState mWhich;
   CGxStateBom    mValue;
-  unsigned int   mStackDepth;
+  unsigned long  mStackDepth;
 };
 
 struct CGxAppRenderState {
@@ -48,7 +48,7 @@ struct CGxAppRenderState {
     mDirty = 0;
   }
 
-  CGxStateBom  mValue;
-  unsigned int mStackDepth;
-  int          mDirty;
+  CGxStateBom mValue;
+  unsigned long mStackDepth;
+  int mDirty;
 };
